+++ CHANGELOG.md
@@ -5,7 +5,8 @@ To see version information, `xsvutils --version`.
 To see compatibility information, `xsvutils help compatibility`.
 
 ## version 0.2.4 (Unreleased)
-- add support subquery of `tee` subcommand
+- add support subquery of `tee` subcommand on v2
+- add `insconst` subcommand on v2
 
 ## version 0.2.3 (2018-03-11)
 - add support CR, CRLF